Back out 1e7050c52cb9 (bug 792850), 776f7325952b (bug 705047), 2e1be0680e70 (bug 733501), 6adef296963c (bug 791775), b8397781f862 (bug 792391) for WinXP hangs and suspicion of breaking Android stacks
UltraBlame original commit: c1e97e2838a75a7114b9f2e8c58ce510d3af434f

+++ toolkit/crashreporter/google-breakpad/src/common/dwarf/bytereader.cc
@@ -123,11 +123,11 @@ uint64 ByteReader::ReadEncodedPointer(const char *buffer,
 
     
     
-    uint64 skew = section_base_ & (AddressSize() - 1);
+    size_t skew = section_base_ & (AddressSize() - 1);
     
-    uint64 offset = skew + (buffer - buffer_base_);
+    off_t offset = skew + (buffer - buffer_base_);
     
-    uint64 aligned = (offset + AddressSize() - 1) & -AddressSize();
+    size_t aligned = (offset + AddressSize() - 1) & -AddressSize();
     
     const char *aligned_buffer = buffer_base_ + (aligned - skew);
     
@@ -156,7 +156,7 @@ uint64 ByteReader::ReadEncodedPointer(const char *buffer,
     case DW_EH_PE_uleb128:
       offset = ReadUnsignedLEB128(buffer, len);
       break;
-
+      
     case DW_EH_PE_udata2:
       offset = ReadTwoBytes(buffer);
       *len = 2;
